move test model to separate namespace
