+++ tests/BasicTest.php
@@ -7,6 +7,7 @@ namespace Atk4\Validate\Tests;
 use Atk4\Data\Model;
 use Atk4\Data\Schema\TestCase;
 use Atk4\Validate\Validator;
+use Atk4\Validate\Tests\Model\ModelTest;
 
 class BasicTest extends TestCase
 {
